# openapi.yaml
openapi: 3.1.0
info:
  title: Sistema Integral de Gestión de Consorcios
  description: |
    API REST para la gestión integral de consorcios. Permite administrar expensas,
    gastos del consorcio, personal y liquidaciones, tareas y presupuestos,
    comunicación interna, reserva de amenities y configuración global. El acceso a
    los recursos está restringido por roles: Administración, Departamentos
    (Inquilinos/Propietarios), Representantes y Super-Admin.
    Fase 3.5 (jun-2026): cuenta corriente por departamento — pagos se asignan FIFO desde movimientos.

    **Multitenant SaaS (Fase 7A, jul-2026)**: la plataforma es multi-tenant con
    jerarquía Super-Admin → Administración → Consorcio → Departamento. Todo
    endpoint operacional (fuera de `/auth/*` y `/me/*`) exige el header
    `X-Consorcio-Id: <id>` para resolver el consorcio activo. El backend valida
    que el usuario del JWT tenga acceso al consorcio pedido; caso contrario
    devuelve **403**. Códigos específicos: `403 cambio_password_requerido`
    (usuario con `must_change_password=True`), `403 administracion_suspendida`
    (login sobre administración inactiva), `403 sin acceso a este consorcio`
    (spoof cross-tenant), `400 X-Consorcio-Id requerido/invalido`.
  version: 1.1.0
servers:
  - url: http://localhost:8000
    description: Servidor local de desarrollo
...
security:
  - bearerAuth: []
paths:
...
  /expensas:
    get:
      tags: [Expensas]
      summary: Listar expensas
      description: |
        Devuelve el historial de expensas. Administración ve todas las expensas
        del consorcio (historial general); los Departamentos ven únicamente las
        asociadas a su propia unidad (aislamiento server-side por el
        `departamento_id` del token). Soporta filtros opcionales por período y
        estado, y paginación mediante `limit` / `offset`.
      operationId: listarExpensas
      parameters:
        - name: periodo
          in: query
          required: false
          description: Filtrar por período (formato YYYY-MM).
          schema:
            type: string
            pattern: '^\d{4}-(0[1-9]|1[0-2])$'
            examples: ["2026-05"]
        - name: departamento_id
          in: query
          required: false
          description: |
            Filtrar por departamento. Sólo efectivo para Administración;
            los usuarios Departamento siempre ven solo sus propias expensas.
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: limit
          in: query
          required: false
          description: Cantidad máxima de resultados a devolver.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          required: false
          description: Cantidad de resultados a saltar (paginación).
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
      responses:
        '200':
          description: Lista de expensas recuperada exitosamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expensa'
        '400':
          $ref: '#/components/responses/PedidoInvalido'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'

    post:
      tags: [Expensas]
      summary: Crear una nueva expensa
      description: |
        Crea una expensa para un período determinado. Solo accesible para
        Administración. Una expensa es única por la combinación
        `(departamento_id, periodo)` — si ya existe se devuelve 409.
      operationId: crearExpensa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpensaCrear'
      responses:
        '201':
          description: Expensa creada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expensa'
        '400':
          $ref: '#/components/responses/PedidoInvalido'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          description: El departamento indicado no existe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Ya existe una expensa para ese departamento y período.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /expensas/{expensa_id}:
    get:
      tags: [Expensas]
      summary: Obtener detalle de una expensa
      description: |
        Devuelve el detalle de una expensa puntual. Los Departamentos solo pueden
        consultar expensas asociadas a su propia unidad.
      operationId: obtenerExpensa
      parameters:
        - name: expensa_id
          in: path
          required: true
          description: Identificador único de la expensa.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Detalle de la expensa recuperado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expensa'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'

    delete:
      tags: [Expensas]
      summary: Eliminar una expensa (admin, sin pagos aplicados)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: expensa_id
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '204':
          description: Expensa eliminada.
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '409':
          description: Conflicto. La expensa tiene pago FIFO aplicado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /expensas/{expensa_id}/pdf:
    get:
      tags: [Expensas]
      summary: Descargar PDF de la boleta
      operationId: descargarPdfExpensa
      security:
        - bearerAuth: []
      parameters:
        - name: expensa_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: PDF de la boleta
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Autenticación mediante token JWT en el header Authorization.
...
  responses:
    PedidoInvalido:
      description: El pedido es inválido o le faltan campos requeridos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NoAutenticado:
      description: No se proporcionó token de autenticación o es inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    AccesoDenegado:
      description: El rol del usuario no tiene permisos para acceder al recurso.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NoEncontrado:
      description: El recurso solicitado no existe.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Descripción legible del error.
          examples: ["No tiene permisos para acceder a este recurso."]

    # ---------------------------- Consorcios (Plan D) ---------------------
...
    ExpensaOut:
      type: object
      properties:
        id: { type: integer }
        departamento_id: { type: integer }
        periodo: { type: string }
        monto_primer_vencimiento: { type: number }
        fecha_primer_vencimiento: { type: string, format: date }
        monto_segundo_vencimiento: { type: number }
        fecha_segundo_vencimiento: { type: string, format: date }
        saldo_anterior: { type: number }
        estado_calculado: { type: string }
        monto_pendiente: { type: number }
        detalle:
          type: array
          items: { $ref: '#/components/schemas/LineaDetalleExpensaOut' }
    Expensa:
      $ref: '#/components/schemas/ExpensaOut'
    ExpensaCrear:
      type: object
      required: [departamento_id, periodo, monto_primer_vencimiento, fecha_primer_vencimiento, monto_segundo_vencimiento, fecha_segundo_vencimiento]
      properties:
        departamento_id: { type: integer }
        periodo: { type: string, pattern: '^\d{4}-(0[1-9]|1[0-2])$' }
        monto_primer_vencimiento: { type: number, exclusiveMinimum: 0 }
        fecha_primer_vencimiento: { type: string, format: date }
        monto_segundo_vencimiento: { type: number, exclusiveMinimum: 0 }
        fecha_segundo_vencimiento: { type: string, format: date }
...
    LineaDetalleExpensaOut:
      type: object
      properties:
        rubro: { type: string }
        clase_prorrateo_id: { type: integer, nullable: true }
        departamento_origen_id: { type: integer, nullable: true }
        concepto: { type: string }
        monto: { type: number }
...